Refactor ReservationSeeder to use AutoMapper for mapping reservation data
